DON'T load resources in static init!!!
