+++ src/main/java/emu/grasscutter/data/DataLoader.java
@@ -114,8 +114,6 @@ public class DataLoader {
         } catch (Exception e) {
             Grasscutter.getLogger().error("An error occurred while trying to check the data folder.", e);
         }
-
-        generateGachaMappings();
     }
 
     private static void checkAndCopyData(String name) {
@@ -129,18 +127,6 @@ public class DataLoader {
 
             Grasscutter.getLogger().debug("Creating default '" + name + "' data");
             FileUtils.copyResource("/defaults/data/" + name, filePath.toString());
-        }
-    }
-
-    private static void generateGachaMappings() {
-        var path = GachaHandler.getGachaMappingsPath();
-        if (!Files.exists(path)) {
-            try {
-                Grasscutter.getLogger().debug("Creating default '" + path + "' data");
-                Tools.createGachaMappings(path);
-            } catch (Exception exception) {
-                Grasscutter.getLogger().warn("Failed to create gacha mappings. \n" + exception);
-            }
         }
     }
 }
